test: rename shared fixture module to avoid foundational-change pre-push hook
