+++ tests/unit/engine/quality/graders/test_llm.py
@@ -16,7 +16,7 @@ from synthorg.engine.quality.verification import (
 from synthorg.engine.workflow.handoff import HandoffArtifact
 from synthorg.providers.enums import FinishReason
 from synthorg.providers.models import CompletionResponse, TokenUsage
-from tests.unit.engine.quality.conftest import (
+from tests.unit.engine.quality.scripted_provider import (
     ScriptedProvider,
     build_tool_call_response,
 )
